Add <leader> to some nmaps
Okay this was crazy. When I opened up Command-T or ctrlp for searching and
pasted in a string that contained the character sequence for one of my nmaps, it
would replace those pasted characters with whatever I had mapped them to. This
was also happening while using feedkeys(). My workaround for now is to just make
sure all nmaps are prefixed with <leader> to decrease the likelihood of hitting
this problem.