Fix issue with null-ls and stylua in non-root directories
The issue was that the 'condition' option for null-ls was being run
before we had any information about the current buffer, so its
"root_has_file" method was simply looking for files relative to the vim
CWD. If we're editing a file like "../myplugin/lua/foo.lua", we can't
search for files (e.g. stylua.toml) relative to the edited file, and so
we may not properly enable/disable formatting.

The fix was to use runtime_condition instead, where we do have access to
the buffer in question. We then cache the result of the dir search to
avoid doing it on each request.