Fix errors when file changes on disk
If the cursor was at the bottom of a file and it changes on disk to be
shorter (e.g. an autoformatter or git operation), we were getting an
error and stacktrace stemming from the LSP highlight references call.
This can be fixed by making sure we pull in file updates from disk.