fix: allow block quote conceals to work with all markdown elements (#1)
With the existing code, the block quote conceals would only work with paragraphs. If there were block quoted code blocks, bulleted lists, etc. this would not work. Also if the paragraph was nested in a header in a block quote this would also not work. I have found this solution to be more universal